python/embedding_service/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
sentence-transformers>=2.3.0
numpy==1.26.0
python-multipart==0.0.6
torch>=2.0.0
openai>=1.1.0
cohere>=4.9.0
httpx>=0.24.0
requests>=2.30.0
huggingface-hub>=0.20.0
tenacity>=8.2.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
cryptography>=41.0.4
prometheus-client>=0.17.1
psutil>=5.9.6
redis>=5.0.1
pytest>=7.4.3
tabulate>=0.9.0
matplotlib>=3.7.0 